Added typing to the http service
